+++ src/ashley/templatetags/custom_tags.py
@@ -3,7 +3,7 @@ from django import template
 from machina.core.db.models import get_model
 from machina.templatetags.forum_tags import forum_list
 
-from ashley.defaults import _FORUM_ROLE_INSTRUCTOR
+from ashley.defaults import _FORUM_ROLE_ADMINISTRATOR, _FORUM_ROLE_INSTRUCTOR
 
 LTIContext = get_model("ashley", "LTIContext")
 
@@ -23,6 +23,23 @@ def is_user_instructor(topic, user):
     # as instructor for this forum
     return any(
         _FORUM_ROLE_INSTRUCTOR in lti.get_user_roles(user)
+        for lti in LTIContext.objects.filter(forum=topic.forum)
+    )
+
+
+@register.filter()
+def is_user_administrator(topic, user):
+    """
+    This will return a boolean indicating if the passed user is admin
+    for the given topic.
+    Usage::
+        {% if topic|is_user_administrator:user %}...{% endif %}
+    """
+    # get the list of lti concerns by this forum, one forum can have multiple LTI Context
+    # if the user is admin in one of them then he is considered
+    # as admin for this forum
+    return any(
+        _FORUM_ROLE_ADMINISTRATOR in lti.get_user_roles(user)
         for lti in LTIContext.objects.filter(forum=topic.forum)
     )